AICc.flexsurvreg {flexsurv}	R Documentation

Second-order Akaike information criterion

Description

Second-order or "corrected" Akaike information criterion, often known as AICc (see, e.g. Burnham and Anderson 2002). This is defined as -2 log-likelihood + (2*p*n)/(n - p -1), whereas the standard AIC is defined as -2 log-likelihood + 2*p, where p is the number of parameters and n is the sample size. The correction is intended to adjust AIC for small-sample bias, hence it only makes a difference to the result for small n.

Usage

## S3 method for class 'flexsurvreg'
AICc(object, cens = TRUE, ...)

## S3 method for class 'flexsurvreg'
AICC(object, cens = TRUE, ...)

Arguments

object
Fitted model returned by flexsurvreg (or flexsurvspline).

cens
Include censored observations in the sample size term (n) used in this calculation. See BIC.flexsurvreg for a discussion of the issues with defining the sample size.

...
Other arguments (currently unused).

Details

This can be spelt either as AICC or AICc.

Value

The second-order AIC of the fitted model.

References

Burnham, K. P., Anderson, D. R. (2002) Model Selection and Multimodel Inference: a practical information-theoretic approach. Second edition. Springer: New York.

See Also

BIC, AIC, BIC.flexsurvreg, nobs.flexsurvreg

[Package flexsurv version 2.3.2 Index]