GenF.orig {flexsurv}	R Documentation

Generalized F distribution (original parameterisation)

Description

Density, distribution function, quantile function and random generation for the generalized F distribution, using the less flexible original parameterisation described by Prentice (1975).

Usage

dgenf.orig(x, mu = 0, sigma = 1, s1, s2, log = FALSE)

pgenf.orig(q, mu = 0, sigma = 1, s1, s2, lower.tail = TRUE, log.p = FALSE)

Hgenf.orig(x, mu = 0, sigma = 1, s1, s2)

hgenf.orig(x, mu = 0, sigma = 1, s1, s2)

qgenf.orig(p, mu = 0, sigma = 1, s1, s2, lower.tail = TRUE, log.p = FALSE)

rgenf.orig(n, mu = 0, sigma = 1, s1, s2)

Arguments

x, q
vector of quantiles.

mu
Vector of location parameters.

sigma
Vector of scale parameters.

s1
Vector of first F shape parameters.

s2
vector of second F shape parameters.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P(X <= x), otherwise, P(X > x).

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

If y ~ F(2*s1, 2*s2), and w = log(y) then x = exp(w*sigma + mu) has the original generalized F distribution with location parameter mu, scale parameter sigma>0 and shape parameters s1>0,s2>0. The probability density function of x is

f(x | mu, sigma, s_1, s_2) = ((s1/s2)^{s1} e^{s1 w}) / (sigma x (1 + s1 e^w/s2) ^ (s1 + s2) B(s1, s2))

where w = (log(x) - mu)/sigma, and B(s1,s2) = Γ(s1)Γ(s2)/Γ(s1+s2) is the beta function.

As s2 -> infinity, the distribution of x tends towards an original generalized gamma distribution with the following parameters:

dgengamma.orig(x, shape=1/sigma, scale=exp(mu) / s1^sigma, k=s1)

See GenGamma.orig for how this includes several other common distributions as special cases.

The alternative parameterisation of the generalized F distribution, originating from Prentice (1975) and given in this package as GenF, is preferred for statistical modelling, since it is more stable as s1 tends to infinity, and includes a further new class of distributions with negative first shape parameter. The original is provided here for the sake of completion and compatibility.

Value

dgenf.orig gives the density, pgenf.orig gives the distribution function, qgenf.orig gives the quantile function, rgenf.orig generates random deviates, Hgenf.orig retuns the cumulative hazard and hgenf.orig the hazard.

Author(s)

Christopher Jackson <chris.jackson@mrc-bsu.cam.ac.uk>

References

R. L. Prentice (1975). Discrimination among some parametric models. Biometrika 62(3):607-614.

See Also

GenF, GenGamma.orig, GenGamma

[Package flexsurv version 2.3.2 Index]