get_basepars {flexsurv}	R Documentation

Evaluate baseline time-to-event distribution parameters given covariate values in a flexsurvmix model

Description

Evaluate baseline time-to-event distribution parameters given covariate values in a flexsurvmix model

Usage

get_basepars(x, newdata, event)

Arguments

x
Fitted model object

newdata
Data frame of alternative covariate values

event
Event

[Package flexsurv version 2.3.2 Index]