lines.flexsurvreg {flexsurv}	R Documentation

Add fitted flexible survival curves to a plot

Description

Add fitted survival (or hazard or cumulative hazard) curves from a flexsurvreg model fit to an existing plot.

Usage

## S3 method for class 'flexsurvreg'
lines(
  x,
  newdata = NULL,
  X = NULL,
  type = "survival",
  t = NULL,
  est = TRUE,
  ci = NULL,
  B = 1000,
  cl = 0.95,
  col = "red",
  lty = 1,
  lwd = 2,
  col.ci = NULL,
  lty.ci = 2,
  lwd.ci = 1,
  ...
)

Arguments

x
Output from flexsurvreg, representing a fitted survival model object.

newdata
Covariate values to produce fitted curves for, as a data frame, as described in plot.flexsurvreg.

X
Covariate values to produce fitted curves for, as a matrix, as described in plot.flexsurvreg.

type
"survival" for survival, "cumhaz" for cumulative hazard, or "hazard" for hazard, as in plot.flexsurvreg.

t
Vector of times to plot fitted values for.

est
Plot fitted curves (TRUE or FALSE.)

ci
Plot confidence intervals for fitted curves.

B
Number of simulations controlling accuracy of confidence intervals, as used in summary.

cl
Width of confidence intervals, by default 0.95 for 95% intervals.

col
Colour of the fitted curve(s).

lty
Line type of the fitted curve(s).

lwd
Line width of the fitted curve(s).

col.ci
Colour of the confidence limits, defaulting to the same as for the fitted curve.

lty.ci
Line type of the confidence limits.

lwd.ci
Line width of the confidence limits, defaulting to the same as for the fitted curve.

...
Other arguments to be passed to the generic plot and lines functions.

Details

Equivalent to plot.flexsurvreg(...,add=TRUE).

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

flexsurvreg

[Package flexsurv version 2.3.2 Index]