model.frame.flexsurvmix {flexsurv}	R Documentation

Model frame from a flexsurvmix object

Description

Returns a list of data frames, with each component containing the data that were used for the model fit for that mixture component.

Usage

## S3 method for class 'flexsurvmix'
model.frame(formula, ...)

Arguments

formula
Fitted model object from flexsurvmix.

...
Further arguments (currently unused).

Value

A list of data frames

[Package flexsurv version 2.3.2 Index]