model.frame.flexsurvreg {flexsurv}	R Documentation

Extract original data from flexsurvreg objects.

Description

Extract the data from a model fitted with flexsurvreg.

Usage

## S3 method for class 'flexsurvreg'
model.frame(formula, ...)

## S3 method for class 'flexsurvreg'
model.matrix(object, par = NULL, ...)

Arguments

formula
A fitted model object, as returned by flexsurvreg.

...
Further arguments (not used).

object
A fitted model object, as returned by flexsurvreg.

par
String naming the parameter whose linear model matrix is desired.

The default value of par=NULL returns a matrix consisting of the model matrices for all models in the object cbinded together, with the intercepts excluded. This is not really a “model matrix” in the usual sense, however, the columns directly correspond to the covariate coefficients in the matrix of estimates from the fitted model.

Value

model.frame returns a data frame with all the original variables used for the model fit.

model.matrix returns a design matrix for a part of the model that includes covariates. The required part is indicated by the "par" argument (see above).

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

flexsurvreg, model.frame, model.matrix.

[Package flexsurv version 2.3.2 Index]