nobs.flexsurvreg {flexsurv}	R Documentation

Number of observations contributing to a fitted flexible survival model

Description

Number of observations contributing to a fitted flexible survival model

Usage

## S3 method for class 'flexsurvreg'
nobs(object, cens = TRUE, ...)

Arguments

object
Output from flexsurvreg or flexsurvspline, representing a fitted survival model object.

cens
Include censored observations in the number. TRUE by default. If FALSE then the number of observed events is returned. See BIC.flexsurvreg for a discussion of the issues with defining the sample size for censored data.

...
Further arguments passed to or from other methods. Currently unused.

Details

By default, this matches the behaviour of the nobs method for survreg objects, including both censored and uncensored observations.

If a weighted flexsurvreg analysis was done, then this function returns the sum of the weights.

Value

This returns the mod$N component of the fitted model object mod. See flexsurvreg, flexsurvspline for full documentation of all components.

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

flexsurvreg, flexsurvspline.

[Package flexsurv version 2.3.2 Index]