Extract model coefficients from fitted flexible survival models

Extract model coefficients from fitted flexible survival models. This presents all parameter estimates, transformed to the real line if necessary. For example, shape or scale parameters, which are constrained to be positive, are returned on the log scale.

Usage

# S3 method for class 'flexsurvreg'
coef(object, ...)

Arguments

object

Output from flexsurvreg or flexsurvspline, representing a fitted survival model object.

...

Further arguments passed to or from other methods. Currently unused.

Value

This returns the mod$res.t[,"est"] component of the fitted model object mod. See flexsurvreg, flexsurvspline for full documentation of all components.

Details

This matches the behaviour of coef.default for standard R model families such as glm, where intercepts in regression models are presented on the same scale as the covariate effects. Note that any parameter in a distribution fitted by flexsurvreg or flexsurvreg may be an intercept in a regression model.

See also

Author

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk